Fix for CWE-1022  CWE-200
`HTML links that open in a new tab or window allow the target page to access the DOM of the origin page using window.opener unless link type noopener or noreferrer is specified. This is a potential security risk.`

In the following example, a JSX element is created that corresponds to an HTML link opening the URL http://example.com in a new tab. Since it does not specify a link type, that page will be able to access the DOM of the origin page.

var link = <a href="http://example.com" target="_blank">Example</a>;

To fix this vulnerability, add a rel attribute:

var link = <a href="http://example.com" target="_blank" rel="noopener noreferrer">Example</a>;